move dashboard content to individual module
